create commission model with logic

+++ db/migrate/20170917131254_create_commissions.rb
@@ -2,7 +2,7 @@ class CreateCommissions < ActiveRecord::Migration[5.1]
   def change
     create_table :commissions do |t|
       t.belongs_to :race, foreign_key: true, index: true
-      t.integer :value
+      t.decimal :value
       t.integer :starts
       t.integer :ends
